+++ src/components/dialog/dialog.vue
@@ -1,15 +1,15 @@
 <template>
   <transition enter-active-class="weui-animate-fade-in" leave-active-class="weui-animate-fade-out">
-    <div v-show="this.value">
+    <div v-show="value">
       <div class="weui-mask"></div>
       <div class="weui-dialog" :class="{ 'weui-skin_android': skin === 'android' }">
         <div class="weui-dialog__hd" v-if="title"><strong class="weui-dialog__title" v-html="title"></strong></div>
         <div class="weui-dialog__bd" v-html="message"></div>
         <div class="weui-dialog__ft">
-          <a class="weui-dialog__btn weui-dialog__btn_default" v-if="showCancelBtn"
-             @click="handleAction('cancel')" v-text="cancelText"></a>
-          <a class="weui-dialog__btn weui-dialog__btn_primary" v-if="showConfirmBtn"
-             @click="handleAction('confirm')" v-text="confirmText"></a>
+          <a class="weui-dialog__btn weui-dialog__btn_default" v-if="showCancelButton"
+             @click="handleAction('cancel')" v-text="cancelButtonText"></a>
+          <a class="weui-dialog__btn weui-dialog__btn_primary" v-if="showConfirmButton"
+             @click="handleAction('confirm')" v-text="confirmButtonText"></a>
         </div>
       </div>
     </div>
@@ -17,11 +17,15 @@
 </template>
 
 <script>
+  import PopupMixin from '../../mixins/popup'
+
   const CONFIRM_TEXT = '确定'
   const CANCEL_TEXT = '取消'
 
   export default {
     name: 'wv-dialog',
+
+    mixins: [PopupMixin],
 
     props: {
       skin: {
@@ -30,37 +34,30 @@
       },
       title: String,
       message: String,
-      confirmText: {
+      confirmButtonText: {
         type: String,
         default: CONFIRM_TEXT
       },
-      cancelText: {
+      cancelButtonText: {
         type: String,
         default: CANCEL_TEXT
       },
-      showConfirmBtn: {
+      showConfirmButton: {
         type: Boolean,
         default: true
       },
-      showCancelBtn: {
+      showCancelButton: {
         type: Boolean,
         default: true
-      }
-    },
-
-    data () {
-      return {
-        value: false
-      }
+      },
+      callback: Function
     },
 
     methods: {
       handleAction (action) {
-        this.value = false
-        if (action === 'confirm') {
-          let callback = this.callback
-          callback(action)
-        }
+        this.$emit('input', false)
+        this.$emit(action)
+        this.callback && this.callback(action)
       }
     }
   }