+++ src/mixins/popup/popup-context.js
@@ -1,0 +1,16 @@
+const PopupContext = {
+  idSeed: 1,
+  zIndex: 2000,
+  instances: {},
+  stack: [],
+
+  plusKeyByOne (key) {
+    return this[key]++
+  },
+
+  get top () {
+    return this.stack[this.stack.length - 1]
+  }
+}
+
+export default PopupContext